refactor: separate block parsers from file parsers
